Successful remote of simple program
Allocates, memcpys, and launches kernels...what else is there?

I had to rip a non trivial amount of code out of the HIP runtime to inspect the
binary and load the kernels properly. The hcc compiler also links against the
system's installed rocm stack no matter what (super annoying) so LD_PRELOAD is a
must to make this work.

I think I should be able to use LD_PRELOAD to intercept any HIP call without
recompiling the extensive number of libraries that we require, which is a huge
boon in that it greatly reduces the source we need to manage.

Signed-off-by: Tyler Hunt <thunt@cs.utexas.edu>