decoupled copy out.
--gpu-op-sep-memcpy to lgm_run is the flag that enables this change.

So copy out turns out to be a little trickier. I can't synchronize with the
kernel launch stream since that would make the copy out time depend on the
completion of the kernels.

My solution is to do this:
Run an copy kernel on the device at the point in the stream where the copy
should happen. This kernel writes the necessary data to a staging buffer on the
GPU and appends a tag.

I keep a list of outstanding d2h memcpys on the client. Every iteration, I copy
the data out of the staging buffer. I look at the tag and if it matches what I
expect (meaning the device copy kernel ran) I mark that d2h memcpy as complete
and start working on the next one.

To be fixed rate we need to add some logic to copy out some static garbage
buffer when there are no outstanding d2h memcpys.

Signed-off-by: Tyler Hunt <thunt@cs.utexas.edu>