My staging buffer logic is pretty simple:
there are N buffers which get used in a ring. as long as there are no more than
N outstanding operations then it is always safe to use the next buffer.

"outstanding operation" means something different for copy in and copy out
operations. Separating their buffers like this allows all mxnet workloads to run
to completion.

Signed-off-by: Tyler Hunt <thunt@cs.utexas.edu>