use pinned buffers for memcpy
Had to solve some weird coherency problems to get it to work. What I really
needed (i think) was __threadfence_system() so that my unsyncronized memcpys
would read the data that I copied out for them. Unfortunately it's not supported
in ROCm 1.8. So I have to launch an extra kernel in the unencrypted case to make
sure a copy on a different stream reads the data.

Signed-off-by: Tyler Hunt <thunt@cs.utexas.edu>